# public/docs/openapi.yaml
openapi: 3.1.0
info:
  title: check-manage AI 能力对外 API
  version: "1.0.0"
  description: |
    覆盖对外 API Key 鉴权的六大 AI 能力家族：AI 批任务、AI 单会话、AI 定时扫描任务触发、
    行操作触发、Prompt 模板、长期记忆。**不含** `open-api.md` 描述的数据集合 CRUD 接口
    （那部分是独立的 `open-api-reference.html`/openapi 文档范围）。

    详细的流程说明、限制、常见问题见各自的 Markdown 文档（本文件只是机器可读的契约，
    人读的叙述性内容请参见）：
    - docs/user-guide/integration/ai-batch-api.md
    - docs/user-guide/integration/ai-session-api.md
    - docs/user-guide/ai/scan-tasks.md（§10）
    - docs/user-guide/data/row-actions.md（§11）
    - docs/user-guide/ai/batch-tasks.md（Prompt 模板 → 对外 API）
    - docs/user-guide/ai/long-term-memory.md（对外 API）

    **鉴权**：所有端点要求请求头 `X-API-Key: <密钥>`，且该密钥必须已绑定用户（在密钥管理页
    新建的密钥默认绑定创建者本人；存量未绑定密钥会被拒绝，需要重新创建）。

    **错误响应**：统一的 JSON 形状 `{"error": "...", "code": "..."}`。`error` 是人读文案
    （AI 批任务/单会话/扫描任务/行操作家族为中文，Prompt 模板/长期记忆家族沿用各自内部路由
    的既有语言，行操作端点的 `error` 是唯一保留中文原文的例外——见 row-actions.md §11 的
    专门说明）；`code` 是新增的机器可判别字段，取值见下方 `ErrorCode` schema。未预期的服务端
    异常也保证返回这个 JSON 形状（而不是 HTML 错误页），`code` 为 `INTERNAL_ERROR`。
  license:
    name: Internal use only
servers:
  - url: https://<host>/api
    description: 生产环境（经 proxy.py 反向代理）
  - url: http://localhost:3002
    description: 开发环境直连后端
security:
  - ApiKeyAuth: []
...
paths:
...
  /v1/ai-scan-tasks/{taskId}:
    parameters:
      - in: path
        name: taskId
        required: true
        schema: { type: string }
    get:
      tags: [ai-scan-tasks]
      summary: 查询单个扫描任务
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScanTask' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/ai-scan-tasks/{taskId}/run-now:
    parameters:
      - in: path
        name: taskId
        required: true
        schema: { type: string }
    post:
      tags: [ai-scan-tasks]
      summary: 立即触发一次扫描（同步认领 + 建批任务，AI 处理异步）
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  triggered: { type: boolean }
                  claimedCount: { type: integer, nullable: true }
                  lastError: { type: string, nullable: true }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /v1/prompt-templates:
    get:
      tags: [prompt-templates]
      summary: 列出当前用户的全部模板
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  templates:
                    type: array
                    items: { $ref: '#/components/schemas/PromptTemplate' }
    post:
      tags: [prompt-templates]
      summary: 新建模板
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, content]
              properties:
                name: { type: string, maxLength: 200 }
                content: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PromptTemplate' }
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/prompt-templates/{templateId}:
    parameters:
      - in: path
        name: templateId
        required: true
        schema: { type: string }
    get:
      tags: [prompt-templates]
      summary: 查询单个模板
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PromptTemplate' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [prompt-templates]
      summary: 修改模板
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, content]
              properties:
                name: { type: string, maxLength: 200 }
                content: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PromptTemplate' }
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [prompt-templates]
      summary: 删除模板
      responses:
        '204':
          description: No Content
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/memories:
    get:
      tags: [memories]
      summary: 列出当前用户的全部长期记忆
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  memories:
                    type: array
                    items: { type: object, description: 原样透传 mem0 返回形状 }
    post:
      tags: [memories]
      summary: 补写一条记忆
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, maxLength: 2000 }
                verbatim: { type: boolean, default: false }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  memories:
                    type: array
                    items: { type: object }
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '409':
          description: 记忆功能未配置/未启用
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  code: { type: string, enum: [MEMORY_UNAVAILABLE] }
  /v1/memories/{memoryId}:
    parameters:
      - in: path
        name: memoryId
        required: true
        schema: { type: string }
        description: mem0 生成的不透明 UUID
    delete:
      tags: [memories]
      summary: 删除一条记忆
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
    ErrorCode:
      type: string
      enum:
        - INVALID_ARGUMENT
        - NOT_FOUND
        - UNAUTHORIZED
        - FORBIDDEN
        - PAYLOAD_TOO_LARGE
        - CONFLICT
        - INTERNAL_ERROR
        - UPSTREAM_UNAVAILABLE
        - MEMORY_UNAVAILABLE
      description: |
        MEMORY_UNAVAILABLE 只出现在 /v1/memories 的 409 响应里；其余七个across 六大家族通用。
        UNAUTHORIZED/FORBIDDEN 在 401（密钥缺失/无效/已停用）场景下不一定携带 code
        字段——那一步鉴权发生在更底层的网关，早于各蓝图自己的错误处理。
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: 人读文案，语言因端点家族而异，见顶层 description }
        code: { $ref: '#/components/schemas/ErrorCode' }
...
    ScanTask:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        enabled: { type: boolean }
        collection: { type: string }
        branchId: { type: string }
        scheduleIntervalMinutes: { type: integer }
        maxRecordsPerScan: { type: integer }
        lastRunAt: { type: string, format: date-time, nullable: true }
        lastScanCount: { type: integer, nullable: true }
        lastError: { type: string, nullable: true }
    PromptTemplate:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        content: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
  responses:
    InvalidArgument:
      description: 参数校验失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: 目标不存在，或存在但不属于本密钥（不区分两种情况，一律 404）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Conflict:
      description: 状态冲突（如对非终态资源做仅限终态的操作）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    InternalError:
      description: 未预期的服务端异常
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }